upgrade DualAxes
