+++ src/AntDesign.Charts.Docs/Demos/Combo/DualLine_Combo_Chart.razor
@@ -1,13 +1,13 @@
 @page "/DualLine_Combo_Chart"
 @inject IJSRuntime JS;
 <Tabs>
-        <TabPane Key="1">
-            <Tab>示例1</Tab>
-            <ChildContent>
-                <DualLine Data="@(new object[] { data1,data2})" Config="config1" />
-            </ChildContent>
-        </TabPane>
-    </Tabs>
+    <TabPane Key="1">
+        <Tab>示例1</Tab>
+        <ChildContent>
+            <DualAxes Data="@(new object[] { data1, data2 })" Config="config1" />
+        </ChildContent>
+    </TabPane>
+</Tabs>
 @code{
 
     #region 示例1
@@ -36,7 +36,7 @@
         new  { year= "1999", count= 20 },
             };
 
-    DualLineConfig config1= new DualLineConfig()
+    DualAxesConfig config1 = new DualAxesConfig()
     {
         Title = new Charts.Title()
         {
@@ -52,9 +52,22 @@
 
         XField = "year",
         YField = new string[] { "value", "count" },
+        GeometryOptions = new object[] {
+        new {
+            Geometry= "line",
+                    Color= "#5B8FF9",
+                },
+
+        new {
+                Geometry= "line",
+                Color= "#5AD8A6",
+            },
+        },
+
     };
 
     #endregion 示例1
 
 }
+
