--- src/AntDesign.Charts/Components/Combo/ColumnLine/ColumnLine.razor
@@ -1,5 +1,0 @@
-﻿@namespace AntDesign.Charts
-
-@inherits ChartComponentBase<ColumnLineConfig>
-
-<div @ref="Ref"></div>
