--- src/AntDesign.Charts/Components/Combo/DualLine/DualLine.razor
@@ -1,5 +1,0 @@
-﻿@namespace AntDesign.Charts
-
-@inherits ChartComponentBase<DualLineConfig>
-
-<div @ref="Ref"></div>
