--- src/AntDesign.Charts/Components/Combo/StackedColumnLine/StackedColumnLine.razor
@@ -1,5 +1,0 @@
-﻿@namespace AntDesign.Charts
-
-@inherits ChartComponentBase<StackedColumnLineConfig>
-
-<div @ref="Ref"></div>
