--- src/AntDesign.Charts/Components/Combo/StackedColumnLine/StackedColumnLine.razor.cs
@@ -1,17 +1,0 @@
-using System.Text.Json.Serialization;
-using Microsoft.AspNetCore.Components;
-using Microsoft.JSInterop;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace AntDesign.Charts
-{
-    public partial class StackedColumnLine : ChartComponentBase<StackedColumnLineConfig>
-    {
-        public StackedColumnLine() : base("StackedColumnLine")
-        {
-
-        }
-    }
-}
-
