+++ src/AntDesign.Charts/Components/Plots/DualAxes/DualAxes.razor
@@ -1,5 +1,5 @@
 ﻿@namespace AntDesign.Charts
 
-@inherits ChartComponentBase<GroupedColumnLineConfig>
+@inherits ChartComponentBase<DualAxesConfig>
 
 <div @ref="Ref"></div>
