+++ src/AntDesign.Charts/Components/Plots/DualAxes/DualAxes.razor.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AntDesign.Charts
+{
+    public partial class DualAxes : ChartComponentBase<DualAxesConfig>
+    {
+        public DualAxes() : base("DualAxes")
+        {
+
+        }
+    }
+}
+
