+++ src/AntDesign.Charts/Components/Plots/DualAxes/DualAxesConfig.cs
@@ -1,25 +1,21 @@
-using OneOf;
+﻿using OneOf;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using System.Text.Json.Serialization;
+using System.Threading.Tasks;
 
 namespace AntDesign.Charts
 {
-    public class GroupedColumnLineConfig : IGroupedColumnLineViewConfig, IPlotConfig
+    public class DualAxesConfig : IDualAxesViewConfig, IPlotConfig
     {
-        [JsonPropertyName("columnGroupField")]
-        public string ColumnGroupField { get; set; }
         [JsonPropertyName("xAxis")]
-        public CatAxis XAxis { get; set; }
+        public ValueCatTimeAxis XAxis { get; set; }
         [JsonPropertyName("tooltip")]
         public object Tooltip { get; set; }
-        [JsonPropertyName("lineSeriesField")]
-        public string LineSeriesField { get; set; }
-        [JsonPropertyName("lineConfig")]
-        public LineConfig LineConfig { get; set; }
-        [JsonPropertyName("columnConfig")]
-        public ColumnConfig ColumnConfig { get; set; }
+        [JsonPropertyName("lineConfigs")]
+        public LineConfig[] LineConfigs { get; set; }
         [JsonPropertyName("title")]
         public Title Title { get; set; }
         [JsonPropertyName("description")]
@@ -32,12 +28,22 @@ namespace AntDesign.Charts
         public string XField { get; set; }
         [JsonPropertyName("yField")]
         public string[] YField { get; set; }
-        [JsonPropertyName("yAxis")]
-        public ComboYAxis YAxis { get; set; }
-        [JsonPropertyName("legend")]
-        public ComboLegendConfig Legend { get; set; }
+        [JsonPropertyName("forceFit")]
+        public bool? ForceFit { get; set; }
+        [JsonPropertyName("width")]
+        public int? Width { get; set; }
         [JsonPropertyName("renderer")]
         public string Renderer { get; set; }
+        [JsonPropertyName("height")]
+        public int? Height { get; set; }
+        [JsonPropertyName("pixelRatio")]
+        public int? PixelRatio { get; set; }
+        [JsonIgnore]
+        public OneOf<string, object> Theme { get; set; }
+        [JsonPropertyName("theme")]
+        public object ThemeMapping => Theme.Value;
+        [JsonPropertyName("localRefresh")]
+        public bool? LocalRefresh { get; set; }
         [JsonIgnore]
         public OneOf<int?, string, int[]> Padding { get; set; }
         [JsonPropertyName("padding")]
@@ -55,10 +61,6 @@ namespace AntDesign.Charts
         [JsonPropertyName("animation")]
         public object AnimationMapping => Animation.Value;
         [JsonIgnore]
-        public OneOf<string, object> Theme { get; set; }
-        [JsonPropertyName("theme")]
-        public object ThemeMapping => Theme.Value;
-        [JsonIgnore]
         public OneOf<string, object> ResponsiveTheme { get; set; }
         [JsonPropertyName("responsiveTheme")]
         public object ResponsiveThemeMapping => ResponsiveTheme.Value;
@@ -72,16 +74,7 @@ namespace AntDesign.Charts
         public ViewConfigDefaultState DefaultState { get; set; }
         [JsonPropertyName("name")]
         public string Name { get; set; }
-        [JsonPropertyName("forceFit")]
-        public bool? ForceFit { get; set; }
-        [JsonPropertyName("width")]
-        public int? Width { get; set; }
-        [JsonPropertyName("height")]
-        public int? Height { get; set; }
-        [JsonPropertyName("pixelRatio")]
-        public int? PixelRatio { get; set; }
-        [JsonPropertyName("localRefresh")]
-        public bool? LocalRefresh { get; set; }
+
         string IViewConfig.YField { get; set; }
         Axis IViewConfig.XAxis { get; set; }
         Axis IViewConfig.YAxis { get; set; }
@@ -92,12 +85,37 @@ namespace AntDesign.Charts
 
         [JsonPropertyName("appendPadding")]
         public int AppendPadding { get; set; }
+
+        [JsonPropertyName("geometryOptions")]
+        public object[] GeometryOptions { get; set; }
+
     }
 
-    public interface IGroupedColumnLineViewConfig : IColumnLineViewConfig
+    public interface IDualAxesViewConfig : IComboViewConfig
     {
-        [JsonPropertyName("columnGroupField")]
-        public string ColumnGroupField { get; set; }
+        [JsonPropertyName("xAxis")]
+        public ValueCatTimeAxis XAxis { get; set; }//OneOf <IValueAxis, ICatAxis, ITimeAxis>
+        [JsonPropertyName("tooltip")]
+        public object Tooltip { get; set; }
+        [JsonPropertyName("lineConfigs")]
+        public LineConfig[] LineConfigs { get; set; }
     }
-}
 
+    public interface IComboViewConfig : IViewConfig
+    {
+        [JsonPropertyName("title")]
+        public Title Title { get; set; }
+        [JsonPropertyName("description")]
+        public Description Description { get; set; }
+        [JsonPropertyName("data")]
+        public object Data { get; set; }//DataItem[][]
+        [JsonPropertyName("meta")]
+        public object Meta { get; set; }//LooseMap<Meta>
+        [JsonPropertyName("xField")]
+        public string XField { get; set; }
+        [JsonPropertyName("yField")]
+        public string[] YField { get; set; }
+    }
+
+
+}
